SQLite Inclusion and Event Creation
Combination of Greg's SQLite Class implementations and event creation.

I have refactored the database code out to their own file(s) and commented out what we are currently not using.
